Moves as much as possible from Models/Packaging, Models/PublishedContent, Models/Entities, Models/Membership
