+++ src/Umbraco.Abstractions/Models/Entities/IMediaEntitySlim.cs
