+++ src/Umbraco.Abstractions/Models/Entities/IMemberEntitySlim.cs
