+++ src/Umbraco.Abstractions/Models/Entities/MediaEntitySlim.cs
