+++ src/Umbraco.Abstractions/Models/Entities/MemberEntitySlim.cs
