+++ src/Umbraco.Abstractions/Models/Packaging/CompiledPackageDocument.cs
