+++ src/Umbraco.Abstractions/Models/Packaging/CompiledPackageFile.cs
