+++ src/Umbraco.Abstractions/Models/Packaging/PreInstallWarnings.cs
