+++ src/Umbraco.Abstractions/Models/PublishedContent/IPublishedPropertyType.cs
