+++ src/Umbraco.Abstractions/Models/PublishedContent/PublishedContentWrapped.cs
