+++ src/Umbraco.Abstractions/Models/PublishedContent/PublishedSearchResult.cs
