+++ src/Umbraco.Abstractions/Models/PublishedContent/UrlMode.cs
