+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -202,11 +202,6 @@
     <Compile Include="Migrations\Upgrade\V_8_0_0\DataTypes\RichTextPreValueMigrator.cs" />
     <Compile Include="Migrations\Upgrade\V_8_0_0\DataTypes\DropDownFlexiblePreValueMigrator.cs" />
     <Compile Include="Migrations\Upgrade\V_8_0_0\MergeDateAndDateTimePropertyEditor.cs" />
-    <Compile Include="Models\ContentEditing\ContentApp.cs" />
-    <Compile Include="Models\ContentEditing\ContentAppBadge.cs" />
-    <Compile Include="Models\ContentEditing\ContentAppBadgeType.cs" />
-    <Compile Include="Models\ContentEditing\IContentAppFactory.cs" />
-    <Compile Include="Models\Entities\EntityExtensions.cs" />
     <Compile Include="Migrations\Upgrade\V_8_0_0\DataTypes\PreValueMigratorBase.cs" />
     <Compile Include="Migrations\Upgrade\V_8_0_0\DataTypes\PreValueDto.cs" />
     <Compile Include="Migrations\Upgrade\V_8_0_0\PropertyEditorsMigrationBase.cs" />
@@ -223,6 +218,7 @@
     <Compile Include="Models\IContentTypeComposition.cs" />
     <Compile Include="Models\IDataType.cs" />
     <Compile Include="Models\Identity\IdentityMapDefinition.cs" />
+    <Compile Include="Models\Identity\IdentityUser.cs" />
     <Compile Include="Models\Identity\UserLoginInfoWrapper.cs" />
     <Compile Include="Models\IMediaType.cs" />
     <Compile Include="Models\IMember.cs" />
@@ -230,7 +226,6 @@
     <Compile Include="Models\Language.cs" />
     <Compile Include="Models\Macro.cs" />
     <Compile Include="Models\MacroProperty.cs" />
-    <Compile Include="Models\Membership\ContentPermissionSet.cs" />
     <Compile Include="Models\Membership\UserGroup.cs" />
     <Compile Include="Models\Membership\UserGroupExtensions.cs" />
     <Compile Include="Models\ObjectTypes.cs" />
@@ -238,32 +233,17 @@
     <Compile Include="Models\PublicAccessEntry.cs" />
     <Compile Include="Migrations\Upgrade\V_8_0_0\FixLanguageIsoCodeLength.cs" />
     <Compile Include="Models\CultureImpact.cs" />
-    <Compile Include="Models\Entities\IMediaEntitySlim.cs" />
-    <Compile Include="Models\Entities\IMemberEntitySlim.cs" />
-    <Compile Include="Models\Entities\MediaEntitySlim.cs" />
-    <Compile Include="Models\Entities\MemberEntitySlim.cs" />
-    <Compile Include="Models\PublishedContent\ILivePublishedModelFactory.cs" />
     <Compile Include="Models\PublishedContent\IndexedArrayItem.cs" />
-    <Compile Include="Models\PublishedContent\IPublishedContent.cs" />
     <Compile Include="Models\PublishedContent\IPublishedContentTypeFactory.cs" />
-    <Compile Include="Models\PublishedContent\IPublishedElement.cs" />
-    <Compile Include="Models\PublishedContent\IPublishedModelFactory.cs" />
-    <Compile Include="Models\PublishedContent\IPublishedProperty.cs" />
-    <Compile Include="Models\PublishedContent\IPublishedValueFallback.cs" />
     <Compile Include="Models\PublishedContent\PublishedContentType.cs" />
-    <Compile Include="Models\PublishedContent\PublishedElementModel.cs" />
-    <Compile Include="Models\PublishedContent\PublishedElementWrapped.cs" />
     <Compile Include="Models\PublishedContent\VariationContextAccessorExtensions.cs" />
     <Compile Include="Models\Template.cs" />
     <Compile Include="Models\TemplateOnDisk.cs" />
-    <Compile Include="Models\PublishedContent\IPublishedContentType.cs" />
-    <Compile Include="Models\PublishedContent\IPublishedPropertyType.cs" />
     <Compile Include="PropertyEditors\IIgnoreUserStartNodesConfig.cs" />
     <Compile Include="PropertyEditors\Validators\DelimitedValueValidator.cs" />
     <Compile Include="PropertyEditors\Validators\RegexValidator.cs" />
     <Compile Include="PropertyEditors\Validators\RequiredValidator.cs" />
     <Compile Include="PublishedContentExtensions.cs" />
-    <Compile Include="Models\PublishedContent\UrlMode.cs" />
     <Compile Include="Persistence\Dtos\PropertyTypeCommonDto.cs" />
     <Compile Include="Persistence\Factories\MacroFactory.cs" />
     <Compile Include="Persistence\Repositories\Implement\ContentTypeCommonRepository.cs" />
@@ -412,12 +392,8 @@
     <Compile Include="Models\ContentTagsExtensions.cs" />
     <Compile Include="Models\ContentTypeBaseExtensions.cs" />
     <Compile Include="Models\DataTypeExtensions.cs" />
-    <Compile Include="Models\Packaging\CompiledPackage.cs" />
-    <Compile Include="Models\Packaging\CompiledPackageDocument.cs" />
-    <Compile Include="Models\Packaging\CompiledPackageFile.cs" />
     <Compile Include="Models\PathValidationExtensions.cs" />
     <Compile Include="Models\PropertyTagsExtensions.cs" />
-    <Compile Include="Models\PublishedContent\NoopPublishedValueFallback.cs" />
     <Compile Include="Models\SimpleContentType.cs" />
     <Compile Include="PackageActions\AllowDoctype.cs" />
     <Compile Include="PackageActions\PublishRootDocument.cs" />
@@ -554,7 +530,6 @@
     <Compile Include="Models\EntityExtensions.cs" />
     <Compile Include="Models\GridValue.cs" />
     <Compile Include="Models\Identity\BackOfficeIdentityUser.cs" />
-    <Compile Include="Models\Identity\IdentityUser.cs" />
     <Compile Include="Models\Media.cs" />
     <Compile Include="Models\MediaExtensions.cs" />
     <Compile Include="Models\MediaType.cs" />
@@ -567,7 +542,6 @@
     <Compile Include="Models\MemberType.cs" />
     <Compile Include="Models\MemberTypePropertyProfileAccess.cs" />
     <Compile Include="Models\Packaging\InstallationSummary.cs" />
-    <Compile Include="Models\Packaging\PreInstallWarnings.cs" />
     <Compile Include="Models\PagedResultOfT.cs" />
     <Compile Include="Models\Property.cs" />
     <Compile Include="Models\PropertyCollection.cs" />
@@ -575,18 +549,11 @@
     <Compile Include="Models\PropertyGroupCollection.cs" />
     <Compile Include="Models\PropertyType.cs" />
     <Compile Include="Models\PropertyTypeCollection.cs" />
-    <Compile Include="Models\PublishedContent\NoopPublishedModelFactory.cs" />
     <Compile Include="Models\PublishedContent\PublishedContentTypeFactory.cs" />
     <Compile Include="Models\PublishedContent\PublishedContentExtensionsForModels.cs" />
     <Compile Include="Models\PublishedContent\PublishedContentModel.cs" />
-    <Compile Include="Models\PublishedContent\PublishedModelAttribute.cs" />
     <Compile Include="Models\PublishedContent\PublishedModelFactory.cs" />
-    <Compile Include="Models\PublishedContent\PublishedContentTypeConverter.cs" />
-    <Compile Include="Models\PublishedContent\PublishedContentWrapped.cs" />
-    <Compile Include="Models\PublishedContent\PublishedPropertyBase.cs" />
     <Compile Include="Models\PublishedContent\PublishedPropertyType.cs" />
-    <Compile Include="Models\PublishedContent\PublishedSearchResult.cs" />
-    <Compile Include="Models\PublishedContent\RawValueProperty.cs" />
     <Compile Include="Persistence\Dtos\AccessDto.cs" />
     <Compile Include="Persistence\Dtos\AccessRuleDto.cs" />
     <Compile Include="Persistence\Dtos\CacheInstructionDto.cs" />
@@ -1095,6 +1062,8 @@
       <Name>Umbraco.Abstractions</Name>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Folder Include="Models\ContentEditing\" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
